requirements.txt
fastapi==0.111.0
uvicorn==0.30.5
SQLAlchemy==2.0.31
alembic==1.13.1
ruff==0.4.8
psycopg2-binary==2.9.9
greenlet==3.0.3
sqlmodel==0.0.18
pydantic_settings==2.8.0
asyncpg==0.29.0
pytest==8.3.4
pytest-asyncio==0.25.3